+++ carts/admin.py
@@ -4,3 +4,6 @@ from .models import *
 
 admin.site.register(Cart)
 admin.site.register(CartItem)
+
+admin.site.register(Order)
+admin.site.register(OrderItem)
