+++ carts/templates/payments/kakaopay.html
@@ -1,0 +1,1 @@
+{{ domain }}
